requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
coverage==7.5.4
dj-database-url==2.2.0
# Django==5.0.7
Django==4.2.9
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
whitehouse==0.3.1
whitenoise==6.7.0
